Specific nodes now overriding corresponding PQNode functions
